+++ extensions/keyboard-shortcut-sequences/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Auto Keyboard Shortcuts Changelog
 
-## [delay for shortcuts and improved ux] - 2025-09-29
+## [delay for shortcuts and improved ux] - 2025-10-14
 
 A delay before a command execution can be added in a sequence.
 
@@ -8,6 +8,7 @@ UX improvements in creating a shortcut sequence:
 
 - Adding and removing shortcut steps got more dynamic and moved to the end of the list.
 - In addition to the method of writing down key codes, a list with special keycodes (volume, brightness, f-keys, aso.) was added. When either of them is used, the other one will be hidden.
+- FN-key as modifier added
 
 ## [Now able to add key codes] - 2023-07-03